backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
alembic>=1.13.0
pymysql>=1.1.0
cryptography>=42.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
aiofiles>=23.2.1
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.4.0
dashscope>=1.15.0
